requirements.txt
python-telegram-bot[job-queue]>=22,<23
Pillow>=10,<12
PyYAML>=6,<7
python-dotenv>=1,<2

